+++ Source/Details/ASGraphicsContext.m
@@ -12,33 +12,10 @@
 
 #import "ASGraphicsContext.h"
 #import <AsyncDisplayKit/ASAssert.h>
+#import <AsyncDisplayKit/ASConfigurationManager.h>
 #import <UIKit/UIGraphics.h>
 #import <UIKit/UIImage.h>
 #import <stdatomic.h>
-
-#pragma mark - Feature Gating
-
-// Two flags that we atomically manipulate to control the feature.
-typedef NS_OPTIONS(uint, ASNoCopyFlags) {
-  ASNoCopyEnabled = 1 << 0,
-  ASNoCopyBlocked = 1 << 1
-};
-static atomic_uint __noCopyFlags;
-
-// Check if it's blocked, and set the enabled flag if not.
-extern BOOL ASEnableNoCopyRendering()
-{
-  ASNoCopyFlags expectedFlags = 0;
-  BOOL enabled = atomic_compare_exchange_strong(&__noCopyFlags, &expectedFlags, ASNoCopyEnabled);
-  ASDisplayNodeCAssert(enabled, @"Can't enable no-copy rendering after first render started.");
-  return enabled;
-}
-
-// Check if it's enabled and set the "blocked" flag either way.
-static BOOL ASNoCopyRenderingBlockAndCheckEnabled() {
-  ASNoCopyFlags oldFlags = atomic_fetch_or(&__noCopyFlags, ASNoCopyBlocked);
-  return (oldFlags & ASNoCopyEnabled) != 0;
-}
 
 #pragma mark - Callbacks
 
@@ -51,7 +28,7 @@ void _ASReleaseCGDataProviderData(__unused void *info, const void *data, __unuse
 
 extern void ASGraphicsBeginImageContextWithOptions(CGSize size, BOOL opaque, CGFloat scale)
 {
-  if (!ASNoCopyRenderingBlockAndCheckEnabled()) {
+  if (!ASActivateExperimentalFeature(ASExperimentalGraphicsContexts)) {
     UIGraphicsBeginImageContextWithOptions(size, opaque, scale);
     return;
   }
@@ -101,7 +78,7 @@ extern void ASGraphicsBeginImageContextWithOptions(CGSize size, BOOL opaque, CGF
 
 extern UIImage * _Nullable ASGraphicsGetImageAndEndCurrentContext()
 {
-  if (!ASNoCopyRenderingBlockAndCheckEnabled()) {
+  if (!ASActivateExperimentalFeature(ASExperimentalGraphicsContexts)) {
     UIImage *image = UIGraphicsGetImageFromCurrentImageContext();
     UIGraphicsEndImageContext();
     return image;
@@ -145,7 +122,7 @@ extern UIImage * _Nullable ASGraphicsGetImageAndEndCurrentContext()
 
 extern void ASGraphicsEndImageContext()
 {
-  if (!ASNoCopyRenderingBlockAndCheckEnabled()) {
+  if (!ASActivateExperimentalFeature(ASExperimentalGraphicsContexts)) {
     UIGraphicsEndImageContext();
     return;
   }